Moved get_info tests into its own class
