+++ .gitignore
@@ -72,3 +72,4 @@ site/
 # Database/testing files
 *.db
 *.zip
+TODO
